requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy
boto3
apache-airflow-providers-apache-spark 
apache-airflow-providers-amazon